refactor modules in scene_creator
