+++ pgdrive/rl/cost.py
@@ -1,4 +1,4 @@
-from pgdrive.scene_creator.ego_vehicle.base_vehicle import BaseVehicle
+from pgdrive.scene_creator.vehicle.base_vehicle import BaseVehicle
 
 
 def pg_cost_function(vehicle: BaseVehicle) -> None:
