+++ pgdrive/rl/observation_type.py
@@ -2,8 +2,8 @@ from abc import ABC
 
 import gym
 import numpy as np
-from pgdrive.scene_creator.ego_vehicle.base_vehicle import BaseVehicle
-from pgdrive.scene_creator.ego_vehicle.vehicle_module.routing_localization import RoutingLocalizationModule
+from pgdrive.scene_creator.vehicle.base_vehicle import BaseVehicle
+from pgdrive.scene_creator.vehicle_module.routing_localization import RoutingLocalizationModule
 from pgdrive.utils.math_utils import clip
 from pgdrive.world.image_buffer import ImageBuffer
