+++ pgdrive/rl/reward.py
@@ -1,5 +1,5 @@
 from pgdrive.utils import clip
-from pgdrive.scene_creator.ego_vehicle.base_vehicle import BaseVehicle
+from pgdrive.scene_creator.vehicle.base_vehicle import BaseVehicle
 
 
 def pg_reward_function(vehicle: BaseVehicle) -> float:
