+++ pgdrive/scene_creator/vehicle/traffic_vehicle.py
