+++ pgdrive/scene_creator/vehicle_module/lidar.py
@@ -7,7 +7,7 @@ from panda3d.core import BitMask32, NodePath
 from pgdrive.constants import BodyName
 from pgdrive.pg_config.cam_mask import CamMask
 from pgdrive.scene_creator.highway_vehicle.behavior import IDMVehicle
-from pgdrive.scene_creator.pg_traffic_vehicle.traffic_vehicle import PGTrafficVehicle
+from pgdrive.scene_creator.vehicle.traffic_vehicle import PGTrafficVehicle
 from pgdrive.utils.asset_loader import AssetLoader
 from pgdrive.utils.coordinates_shift import panda_position
 from pgdrive.world.pg_physics_world import PGPhysicsWorld
