+++ pgdrive/scene_creator/vehicle_module/mini_map.py
