+++ pgdrive/scene_creator/vehicle_module/rgb_camera.py
