+++ pgdrive/tests/test_pgdrive_env/test_navigation.py
@@ -1,5 +1,5 @@
 from pgdrive.envs.pgdrive_env import PGDriveEnv
-from pgdrive.scene_creator.ego_vehicle.vehicle_module.PID_controller import PIDController
+from pgdrive.scene_creator.vehicle_module.PID_controller import PIDController
 from pgdrive.scene_creator.map import Map, MapGenerateMethod
