+++ pgdrive/world/screen_shot_camera.py
@@ -1,7 +1,7 @@
 from panda3d.core import Vec3
 
 from pgdrive.pg_config.cam_mask import CamMask
-from pgdrive.scene_creator.ego_vehicle.vehicle_module.mini_map import MiniMap
+from pgdrive.scene_creator.vehicle_module import MiniMap
 
 
 class ScreenShotCam(MiniMap):
