+++ pgdrive/world/top_down_observation/top_down_observation.py
@@ -6,7 +6,7 @@ import gym
 import numpy as np
 from pgdrive.constants import Decoration, DEFAULT_AGENT, PG_EDITION
 from pgdrive.rl.observation_type import ObservationType
-from pgdrive.scene_creator.ego_vehicle.base_vehicle import BaseVehicle
+from pgdrive.scene_creator.vehicle.base_vehicle import BaseVehicle
 from pgdrive.utils import import_pygame
 from pgdrive.world.top_down_observation.top_down_obs_impl import WorldSurface, ObservationWindow, COLOR_BLACK, \
     VehicleGraphics, LaneGraphics
